Add idmr_core module with clean API for IDC estimation
- Add idmr_core package with simulation (DGP configs), models (IDCEstimator),
  and data modules providing a clean interface over legacy code
- Add pyproject.toml with hatchling build system and uv dependency management
- Add sanity check script comparing IDC vs MLE on toy data
- Fix dtype mismatch in classesv2.py (C, V tensors now cast to float64)
- Update README with Quickstart section for new core API
- Update .gitignore for AGENTS.md and .uv-cache/

+++ classesv2.py
@@ -22,8 +22,8 @@ class nll_l_cvm(nn.Module):
 
     def __init__(self, data_):
         super().__init__()
-        self.C = torch.from_numpy(data_.C) #(n,d)
-        self.V = torch.from_numpy(data_.V) #(n,p)
+        self.C = torch.from_numpy(data_.C).to(torch.float64) #(n,d)
+        self.V = torch.from_numpy(data_.V).to(torch.float64) #(n,p)
         self.m = torch.from_numpy(data_.m).to(torch.float64) #(m,)
         
         self.n = self.C.shape[0]
